+++ crates/ui/main/src/core/views.rs
@@ -1,2 +1,3 @@
+pub mod cache_page;
 pub mod gallery;
 pub mod setup;